Restart + Player color change base functionality
I deleted everything and restarted the project because the low-level networking API is too hard to use right now. I am going to be using a more high-level approach.

Right now a player can host and another player can join the host. The player can walk it's cloud sprite around and change it's color by standing on the circle and clicking on a color.