Fixed the progressbar
The I don't know exactly what the problem was. But it had something to do with the server not properly spawning the roundmanager and the progressbar on nonlocal clients. https://mirror-networking.com/docs/Articles/Guides/GameObjects/SceneObjects.html

It was fixed after I changed the way it switched player objects in XNetworkManager:OnServerSceneChanged(string)